backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
pandas>=2.0.0
scikit-learn>=1.3.0
underthesea>=6.8.0
django-filter>=24.2
Faker>=24.0.0
